+++ test/gui/test_multiworld_client.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import datetime
 import uuid
 from typing import TYPE_CHECKING
 from unittest.mock import AsyncMock, MagicMock, call
@@ -15,7 +14,7 @@ from randovania.interface_common.players_configuration import INVALID_UUID
 from randovania.interface_common.world_database import WorldData, WorldDatabase, WorldServerData
 from randovania.network_common import error
 from randovania.network_common.game_connection_status import GameConnectionStatus
-from randovania.network_common.multiplayer_session import MultiplayerSessionListEntry, MultiplayerUser, MultiplayerWorld
+from randovania.network_common.multiplayer_session import MultiplayerUser, MultiplayerWorld
 from randovania.network_common.session_state import MultiplayerSessionState
 from randovania.network_common.world_sync import (
     ServerSyncRequest,
@@ -219,12 +218,6 @@ async def test_server_sync(client, mocker: MockerFixture):
     uid_2 = uuid.UUID("00000000-0000-1111-0000-000000000000")
     uid_3 = uuid.UUID("000000000000-0000-0000-0000-11111111")
 
-    w1_session = MultiplayerSessionListEntry(
-        id=567, name="The Session", has_password=False, state=MultiplayerSessionState.IN_PROGRESS,
-        num_players=5, creator="Not You", creation_date=datetime.datetime(2019, 1, 3, 2, 50,
-                                                                          tzinfo=datetime.UTC),
-        is_user_in_session=False,
-    )
 
     request = ServerSyncRequest(worlds=frozendict({
         uid_1: ServerWorldSync(
@@ -258,7 +251,8 @@ async def test_server_sync(client, mocker: MockerFixture):
             worlds=frozendict({
                 uid_1: ServerWorldResponse(
                     world_name="World 1",
-                    session=w1_session,
+                    session_id=567,
+                    session_name="The Session",
                 ),
             }),
             errors=frozendict({